backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
motor==3.6.0
pymongo>=4.9,<4.10
langchain==0.3.13
langchain-groq==0.2.4
langchain-community==0.3.13
chromadb==0.4.24
pypdf==4.2.0
python-docx==1.1.2
tiktoken==0.7.0
pydantic==2.10.4
pydantic-settings==2.7.0
aiofiles==24.1.0
numexpr==2.10.2
pytest==8.2.2
